requirements.txt
requests
python-dotenv
sendgrid
pytest